hello_world/requirements.txt
requests
google-api-python-client==2.13.0
google-auth==1.33.0
boto3==1.18.5
slack_sdk==3.10.1

